Updated models.py
when migrating run the python manage.py makemigrations and then python manage.py migrate. If  you are an admin, go on the admin site for our app and click +models you will be able to fill it out. This is version one of the models.py I will keep fixing it